Rename git-flip into git-mini
It was weird to use a verb as the subcommand since it means users typed
two verbs like "git flip sync". For consistency with other git commands
like "git subtree", we use a noun instead so that users run "git mini
sync". The "mini" name also means there is one less thing to explain to
users because we already had a concept of "mini-repo", but nothing
related to "flipping".